Integrating the system

+++ src/main/kotlin/taxiapp/farecalculator/dto/RideFareTO.kt
@@ -6,6 +6,6 @@ import org.springframework.http.HttpStatus
 data class RideFareTO(
     val distance: Int,
     val duration: Long,
-    val amount: Double,
-    val driverProfit: Double
+    val amount: Int,
+    val driverProfit: Int
 )
